requirements.txt
streamlit==1.32.0
google-generativeai==0.8.3
torch==2.2.2
transformers==4.39.3
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
spacy==3.7.4
networkx==3.2.1
streamlit-agraph==0.0.45
ioc-finder==7.1.0
python-dotenv==1.0.1
torchcrf==1.1.0
# Document upload support (PDF and Word file processing)
PyMuPDF>=1.23.0
python-docx>=1.1.0
